requirements.txt
http://nodeload.github.com/django/django/tarball/master#egg=django
http://nodeload.github.com/t0ster/djaml/tarball/master#egg=djaml
psycopg2
django-nose
http://nodeload.github.com/django-debug-toolbar/django-debug-toolbar/tarball/master#egg=django-debug-toolbar
http://nodeload.github.com/jessemiller/HamlPy/tarball/master#egg=HamlPy
mechanize
gunicorn
tweepy
http://nodeload.github.com/gdoermann/django-chargify/tarball/master#egg=django-chargify
django-debug-template
http://nodeload.github.com/t0ster/django-bootstrap-form/tarball/master#egg=django-bootstrap-form
pygeocoder